% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plus.R
\name{op-plus-paste}
\alias{op-plus-paste}
\alias{+}
\title{Plus operator}
\usage{
x + y
}
\arguments{
\item{x, y}{numeric or complex vectors or objects which can be coerced to
such, or other objects for which methods have been written, or a combination
of numeric and character vectors.}
}
\value{
Same as \link[base]{+}, except if \code{e1} and \code{e2} are two character
vectors or a combination of a character vector and a numeric vector, in this
case the result is \code{paste0(x, y)}.
}
\description{
Behaves like normal \link[base]{+} except for character
vectors combinations of character and numeric vectors, which are pasted
together.
}
\examples{
"abc" + "def"
"abc" + 1.234
1.234 + "abc"
"abc" + 2L
2L + "abc"
}
